Fix one wire driver

+++ Inc/ds2781.h
@@ -23,13 +23,13 @@ inline uint16_t DS2781_read16bitReg(DS2781 ds, uint8_t addr);
 inline uint32_t DS2781_read32bitReg(DS2781 ds, uint8_t addr);
 inline uint64_t DS2781_read64bitReg(DS2781 ds, uint8_t addr);
 
-uint16_t DS2781_readVoltage (DS2781 ds);
-uint16_t DS2781_readCurrent (DS2781 ds);
+double DS2781_readVoltage (DS2781 ds);
+double DS2781_readCurrent (DS2781 ds);
 
-uint32_t DS2781_readAccumulatedCurrent (DS2781 ds);
+double DS2781_readAccumulatedCurrent (DS2781 ds);
 void DS2781_resetAccumulatedCurrent (DS2781 ds);
 uint64_t DS2781_readNetAddress (DS2781 ds);
-uint16_t DS2781_readTemperature (DS2781 ds);
+double DS2781_readTemperature (DS2781 ds);
 
 uint8_t DS2781_readCurrentOffset (DS2781 ds);
 void  DS2781_writeCurrentOffset (DS2781 ds, uint8_t offset);
